Fix and enhance llama-stack models tests (#968)
* fix: fix and enhance models tests


* feat:  update model info type and enhance model tests

- Changed `embedding_dimension` type from `int` to `float` in `ModelInfo`.
- Improved test coverage for model listing and retrieval, ensuring compatibility
with OpenAI SDK structure.
- Added assertions to validate model attributes and metadata.
- Enhanced error handling for non-existent models.


* fix: addess enhancements proposed by CodeRabbit


---------

+++ tests/llama_stack/constants.py
@@ -26,7 +26,7 @@ class ModelInfo(NamedTuple):
 
     model_id: str
     embedding_model: Model
-    embedding_dimension: int
+    embedding_dimension: float  # API returns float (e.g., 768.0) despite being conceptually an integer
 
 
 LLS_CORE_POD_FILTER: str = "app=llama-stack"
